requirements.txt
langchain-openai
langchain-community
langchain
langchain-pinecone
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pymongo
pydantic>=2.0.0
python-dotenv
motor
streamlit
pinecone
PyPDF2
whisper
whisper-openai